CEMS-1866 modify hot render request objects. One for creating the request and the request object itself
